+++ scripts/operation/acelerografo/libraries/lector_json.h
@@ -6,9 +6,9 @@ struct datos_config
     char id[10];
     char fuente_reloj[10];
     char deteccion_eventos[10];
-    char archivos_temporales[60];
-    char registro_continuo[60];
-    char eventos_detectados[60];
+    char archivos_temporales[100];
+    char registro_continuo[100];
+    char eventos_detectados[100];
 };
 
 struct datos_config *compilar_json(const char *filename);