+++ scripts/task/comprobar.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+# Definir la raíz del proyecto local
+PROJECT_LOCAL_ROOT="/home/rsa/projects/acelerografo-rsa"
+
+echo " "
+$PROJECT_LOCAL_ROOT/scripts/acelerografo/ejecutables/comprobar_registro
+echo " "
+
+exit 0