+++ scripts/task/resetmaster.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+# Definir la raíz del proyecto local
+PROJECT_LOCAL_ROOT="/home/rsa/projects/acelerografo-rsa"
+
+sudo $PROJECT_LOCAL_ROOT/scripts/acelerografo/ejecutables/reset_master