+++ task-scripts/comprobar.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-echo " "
-/home/rsa/ejecutables/comprobarregistro
-echo " "
-
-exit 0